requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Document Parsing
pdfplumber==0.11.4
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3

# Image Processing
Pillow==11.1.0

# LLM Integration (Groq + Ollama - both OpenAI-compatible)
openai==1.58.1
httpx==0.28.1

# File Handling
python-magic-bin==0.4.14
aiofiles==24.1.0

# Utilities
uuid6==2024.7.10

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# PageIndex (will be installed from git)
# pip install git+https://github.com/VectifyAI/PageIndex.git
